# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "🎶 MusicBoxAPI",
        "description": "API RESTful pour la gestion des artistes, albums et chansons.\n *                   Cette API permet de gérer les ressources principales :\n *                   - Artistes\n *                   - Albums\n *                   - Chansons\n *                   Elle inclut des fonctionnalités de recherche, pagination et filtres.",
        "contact": {
            "name": "MusicBox API Team",
            "email": "support@musicboxapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Lister les artistes",
                "description": "Retourne la liste paginée des artistes",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des artistes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Créer un nouvel artiste",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Eminem"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rap"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "USA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Artiste créé avec succès"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Afficher un artiste par ID",
                "operationId": "d0b68c06b8bc310fbe24c6d3ea80dac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artiste trouvé"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Mettre à jour un artiste",
                "operationId": "5a5632ea389f9a05336f8d035f5a44c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Slim Shady"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Hip-Hop"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "USA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Artiste mis à jour"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Supprimer un artiste",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Artiste supprimé avec succès"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            }
        },
        "/api/artists/genre/{genre}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Lister les artistes par genre",
                "operationId": "2a7e6b27348da90a55d74ed9984c03f6",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "description": "Genre musical",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Rap"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des artistes"
                    },
                    "404": {
                        "description": "Aucun artiste trouvé pour ce genre"
                    }
                }
            }
        },
        "/api/artists/name/{name}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Trouver un artiste par nom",
                "operationId": "51857c1910829c0668a41c6f5726bf4f",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nom de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Eminem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artiste trouvé"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            }
        },
        "/api/artists/{id}/details": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Afficher un artiste avec ses albums",
                "operationId": "5d27e679115aff47d01483df92ca261b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'artiste",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'artiste et ses albums"
                    },
                    "404": {
                        "description": "Artiste non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "API Token",
                "scheme": "bearer"
            }
        }
    },
...
}